Move  LogClient.GetEntries to its own file.
